requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
